chore: (Models) Rename cell.schema.ts and section.schema.ts
